+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ==================
 - Remove hardcoded api-key
 - Fix 1 GP result bug
+- Search `Address1` and `Address2` fields in typeahead
+- Improve display of typeahead results
 
 0.4.0 / 2018-07-30
 ==================